Split lib.rs
Split lib.rs into config.rs and parser.rs to clearly separate JSON/TOML
parsers and the Config implementation.  JSON and TOML parsers are kept
in the same file to easy synchronization and review.
